Optimized Xsens buffering
Previously had Xsens buffer in non-cacheable region for DMA cache coherence, now we manually invalidate D-cache on UART DMA receive. This allows us to take advantage of D-cache  for the large buffer. Also fixed some compilation issues from previous commits.

+++ TelemHost_Firmware_FE13/Core/Src/serial.c
@@ -1,6 +1,6 @@
 #include "serial.h"
 
-void Serial_Init(Serial_t* serial, UART_HandleTypeDef* h_uart, uint8_t* buf, uint8_t len)
+void Serial_Init(Serial_t* serial, UART_HandleTypeDef* h_uart, uint8_t* buf, uint16_t len)
 {
 	serial->h_uart = h_uart;
 	serial->buf = buf;
